reorganize includes to highlight the important ones
