+++ cmake_bindings/bdsg/is_single_stranded.cpp
@@ -1,4 +1,4 @@
-#include <bdsg/is_single_stranded.hpp>
+#include <bdsg/internal/is_single_stranded.hpp>
 #include <functional>
 #include <handlegraph/handle_graph.hpp>
 #include <handlegraph/types.hpp>
@@ -15,7 +15,7 @@
 #include <pybind11/stl_bind.h>
 #include <pybind11/functional.h>
 #include <fstream>
-#include <bdsg/binder_hook_compile.hpp>
+#include <bdsg/internal/binder_hook_compile.hpp>
 
 
 #ifndef BINDER_PYBIND11_TYPE_CASTER
