+++ cmake_bindings/bdsg/packed_path_position_overlays.cpp
@@ -1,4 +1,4 @@
-#include <bdsg/packed_path_position_overlays.hpp>
+#include <bdsg/overlays/packed_path_position_overlays.hpp>
 #include <functional>
 #include <handlegraph/handle_graph.hpp>
 #include <handlegraph/path_handle_graph.hpp>
@@ -17,7 +17,7 @@
 #include <pybind11/stl_bind.h>
 #include <pybind11/functional.h>
 #include <fstream>
-#include <bdsg/binder_hook_compile.hpp>
+#include <bdsg/internal/binder_hook_compile.hpp>
 
 
 #ifndef BINDER_PYBIND11_TYPE_CASTER
