+++ cmake_bindings/bdsg/path_position_overlays.cpp
@@ -1,4 +1,4 @@
-#include <bdsg/path_position_overlays.hpp>
+#include <bdsg/overlays/path_position_overlays.hpp>
 #include <functional>
 #include <handlegraph/handle_graph.hpp>
 #include <handlegraph/mutable_path_deletable_handle_graph.hpp>
@@ -19,7 +19,7 @@
 #include <pybind11/stl_bind.h>
 #include <pybind11/functional.h>
 #include <fstream>
-#include <bdsg/binder_hook_compile.hpp>
+#include <bdsg/internal/binder_hook_compile.hpp>
 
 
 #ifndef BINDER_PYBIND11_TYPE_CASTER
