+++ cmake_bindings/bdsg/split_strand_graph.cpp
@@ -1,5 +1,5 @@
 #include <bdsg/packed_graph.hpp>
-#include <bdsg/split_strand_graph.hpp>
+#include <bdsg/internal/split_strand_graph.hpp>
 #include <functional>
 #include <handlegraph/handle_graph.hpp>
 #include <handlegraph/path_handle_graph.hpp>
@@ -22,7 +22,7 @@
 #include <pybind11/stl_bind.h>
 #include <pybind11/functional.h>
 #include <fstream>
-#include <bdsg/binder_hook_compile.hpp>
+#include <bdsg/internal/binder_hook_compile.hpp>
 
 
 #ifndef BINDER_PYBIND11_TYPE_CASTER
