+++ cmake_bindings/bdsg/utility.cpp
@@ -1,5 +1,5 @@
 #include <bdsg/hash_graph.hpp>
-#include <bdsg/utility.hpp>
+#include <bdsg/internal/utility.hpp>
 #include <functional>
 #include <handlegraph/handle_graph.hpp>
 #include <handlegraph/path_handle_graph.hpp>
@@ -21,7 +21,7 @@
 #include <pybind11/stl_bind.h>
 #include <pybind11/functional.h>
 #include <fstream>
-#include <bdsg/binder_hook_compile.hpp>
+#include <bdsg/internal/binder_hook_compile.hpp>
 
 
 #ifndef BINDER_PYBIND11_TYPE_CASTER
