+++ cmake_bindings/std/bdsg/binder_hook_bind.cpp
@@ -28,7 +28,7 @@
 #include <pybind11/stl_bind.h>
 #include <pybind11/functional.h>
 #include <fstream>
-#include <bdsg/binder_hook_compile.hpp>
+#include <bdsg/internal/binder_hook_compile.hpp>
 
 
 #ifndef BINDER_PYBIND11_TYPE_CASTER
@@ -40,16 +40,16 @@
 
 void bind_std_bdsg_binder_hook_bind(std::function< pybind11::module &(std::string const &namespace_) > &M)
 {
-	// std::vector file:bdsg/binder_hook_bind.hpp line:26
+	// std::vector file:bdsg/internal/binder_hook_bind.hpp line:26
 	binder::vector_binder<unsigned long,std::allocator<unsigned long>>(M("std"), "unsigned_long", "std_allocator_unsigned_long_t");
 
-	// std::vector file:bdsg/binder_hook_bind.hpp line:27
+	// std::vector file:bdsg/internal/binder_hook_bind.hpp line:27
 	binder::vector_binder<handlegraph::handle_t,std::allocator<handlegraph::handle_t>>(M("std"), "handlegraph_handle_t", "std_allocator_handlegraph_handle_t_t");
 
-	// std::vector file:bdsg/binder_hook_bind.hpp line:28
+	// std::vector file:bdsg/internal/binder_hook_bind.hpp line:28
 	binder::vector_binder<handlegraph::path_handle_t,std::allocator<handlegraph::path_handle_t>>(M("std"), "handlegraph_path_handle_t", "std_allocator_handlegraph_path_handle_t_t");
 
-	// std::vector file:bdsg/binder_hook_bind.hpp line:29
+	// std::vector file:bdsg/internal/binder_hook_bind.hpp line:29
 	binder::vector_binder<handlegraph::step_handle_t,std::allocator<handlegraph::step_handle_t>>(M("std"), "handlegraph_step_handle_t", "std_allocator_handlegraph_step_handle_t_t");
 
 }
