+++ include/bdsg/internal/endianness.hpp
