+++ include/bdsg/internal/is_single_stranded.hpp
