+++ include/bdsg/internal/utility.hpp
