+++ include/bdsg/internal/wang_hash.hpp
