+++ include/bdsg/packed_graph.hpp
@@ -13,12 +13,12 @@
 #include <handlegraph/mutable_path_deletable_handle_graph.hpp>
 #include <handlegraph/serializable_handle_graph.hpp>
 
-#include "bdsg/packed_structs.hpp"
-#include "bdsg/split_strand_graph.hpp"
-#include "bdsg/hash_map.hpp"
-#include "bdsg/utility.hpp"
+#include "bdsg/internal/packed_structs.hpp"
+#include "bdsg/internal/split_strand_graph.hpp"
+#include "bdsg/internal/hash_map.hpp"
+#include "bdsg/internal/utility.hpp"
 
-#include "bdsg/eades_algorithm.hpp"
+#include "bdsg/internal/eades_algorithm.hpp"
 
 
 namespace bdsg {
